[ini.h] Fix -Wsign-compare warnings

+++ ini.h
@@ -792,7 +792,7 @@ int ini_section_add( ini_t* ini, char const* name, int length )
             }
 
         ini->sections[ ini->section_count ].name_large = 0;
-        if( length + 1 >= sizeof( ini->sections[ 0 ].name ) )
+        if( (size_t) length + 1 >= sizeof( ini->sections[ 0 ].name ) )
             {
             ini->sections[ ini->section_count ].name_large = (char*) INI_MALLOC( ini->memctx, (size_t) length + 1 );
             INI_MEMCPY( ini->sections[ ini->section_count ].name_large, name, (size_t) length );
@@ -834,7 +834,7 @@ void ini_property_add( ini_t* ini, int section, char const* name, int name_lengt
         ini->properties[ ini->property_count ].name_large = 0;
         ini->properties[ ini->property_count ].value_large = 0;
 
-        if( name_length + 1 >= sizeof( ini->properties[ 0 ].name ) )
+        if( (size_t) name_length + 1 >= sizeof( ini->properties[ 0 ].name ) )
             {
             ini->properties[ ini->property_count ].name_large = (char*) INI_MALLOC( ini->memctx, (size_t) name_length + 1 );
             INI_MEMCPY( ini->properties[ ini->property_count ].name_large, name, (size_t) name_length );
@@ -846,7 +846,7 @@ void ini_property_add( ini_t* ini, int section, char const* name, int name_lengt
             ini->properties[ ini->property_count ].name[ name_length ] = '\0';
             }
 
-        if( value_length + 1 >= sizeof( ini->properties[ 0 ].value ) )
+        if( (size_t) value_length + 1 >= sizeof( ini->properties[ 0 ].value ) )
             {
             ini->properties[ ini->property_count ].value_large = (char*) INI_MALLOC( ini->memctx, (size_t) value_length + 1 );
             INI_MEMCPY( ini->properties[ ini->property_count ].value_large, value, (size_t) value_length );
@@ -917,7 +917,7 @@ void ini_section_name_set( ini_t* ini, int section, char const* name, int length
         if( ini->sections[ section ].name_large ) INI_FREE( ini->memctx, ini->sections[ section ].name_large );
         ini->sections[ section ].name_large = 0;
         
-        if( length + 1 >= sizeof( ini->sections[ 0 ].name ) )
+        if( (size_t) length + 1 >= sizeof( ini->sections[ 0 ].name ) )
             {
             ini->sections[ section ].name_large = (char*) INI_MALLOC( ini->memctx, (size_t) length + 1 );
             INI_MEMCPY( ini->sections[ section ].name_large, name, (size_t) length );
@@ -945,7 +945,7 @@ void ini_property_name_set( ini_t* ini, int section, int property, char const* n
             if( ini->properties[ p ].name_large ) INI_FREE( ini->memctx, ini->properties[ p ].name_large );
             ini->properties[ ini->property_count ].name_large = 0;
 
-            if( length + 1 >= sizeof( ini->properties[ 0 ].name ) )
+            if( (size_t) length + 1 >= sizeof( ini->properties[ 0 ].name ) )
                 {
                 ini->properties[ p ].name_large = (char*) INI_MALLOC( ini->memctx, (size_t) length + 1 );
                 INI_MEMCPY( ini->properties[ p ].name_large, name, (size_t) length );
@@ -974,7 +974,7 @@ void ini_property_value_set( ini_t* ini, int section, int property, char const* 
             if( ini->properties[ p ].value_large ) INI_FREE( ini->memctx, ini->properties[ p ].value_large );
             ini->properties[ ini->property_count ].value_large = 0;
 
-            if( length + 1 >= sizeof( ini->properties[ 0 ].value ) )
+            if( (size_t) length + 1 >= sizeof( ini->properties[ 0 ].value ) )
                 {
                 ini->properties[ p ].value_large = (char*) INI_MALLOC( ini->memctx, (size_t) length + 1 );
                 INI_MEMCPY( ini->properties[ p ].value_large, value, (size_t) length );
